fix: refactor definitions
