+++ src/lib/queries/getProcessDefinitions.ts
@@ -1,19 +1,36 @@
-import { ContentDefinition, ProcessDefinitionData } from 'fluro'
-
 import { client } from '../fluro'
 
-export interface GetProcessDefinitions {
-  [definitionName: string]: ContentDefinition<ProcessDefinitionData>
+export type DefinitionName = 'contact' | 'process'
+
+interface ProcessDefinition {
+  _id: string
+  definitionName: string
+  data: { states: { title: string; key: string } }
+  firstLine: string
+  plural: string
+  title: string
 }
 
-export async function getProcessDefinitions(): Promise<GetProcessDefinitions> {
-  return client.types.retrieve(['process']).then((data) => {
+interface Definition {
+  definitionName: string
+  definitions: ProcessDefinition[]
+}
+
+export interface GetProcessDefinitions {
+  [definitionName: string]: ProcessDefinition
+}
+
+export function getProcessDefinitions(): () => Promise<GetProcessDefinitions> {
+  return async (): Promise<GetProcessDefinitions> => {
+    const data = await client.types.retrieve<Definition>(['process'])
     const definitions: {
-      [key: string]: ContentDefinition<ProcessDefinitionData>
+      [key: string]: ProcessDefinition
     } = {}
+
     data[0].definitions.forEach((definition): void => {
       definitions[definition.definitionName] = definition
     })
+
     return definitions
-  })
+  }
 }
